#64: implemented genmodel based type provider.
Refactored EMFTypeProvider to provide a much more readable and
extendable code. Resolution logic: if the main resolution of a
variable's type fails, the GenModelBasedTypeProvider tries to find a
GenClass for the appropriate EClassifier. If exist, the fully
qualified name of the GenClass interface is used to find the Java class.